Move code into settings class, fixes for CS & docs
